Move helper to rule folder
