+++ tests/Unit/Rule/StrAlphaUnderscoreTest.php
@@ -3,22 +3,22 @@
 namespace Tests\Unit\Rule;
 
 use App\Rules\StrAlphaUnderscore;
-use Illuminate\Validation\Validator;
 use Tests\TestCase;
 
 class StrAlphaUnderscoreTest extends TestCase
 {
+    use Helper;
+
     /**
      * @test
      * @group u-rule
      */
     public function strAlphaUnderscorePass()
     {
-        $trans = $this->getIlluminateArrayTranslator();
-        $validator = new Validator($trans, ['foo' => 'foo_bar'], ['foo' => new StrAlphaUnderscore]);
+        $val = $this->validator(['foo' => 'foo_bar'], ['foo' => new StrAlphaUnderscore]);
 
-        $this->assertTrue($validator->passes());
-        $this->assertSame([], $validator->messages()->messages());
+        $this->assertTrue($val->passes());
+        $this->assertSame([], $val->messages()->messages());
     }
 
     /**
@@ -27,17 +27,16 @@ class StrAlphaUnderscoreTest extends TestCase
      */
     public function strAlphaUnderscoreFail()
     {
-        $trans = $this->getIlluminateArrayTranslator();
-        $validator = new Validator($trans, ['foo' => 'foo-bar'], ['foo' => new StrAlphaUnderscore]);
-        $this->assertTrue($validator->fails());
+        $val = $this->validator(['foo' => 'foo-bar'], ['foo' => new StrAlphaUnderscore]);
+        $this->assertTrue($val->fails());
 
-        $validator = new Validator($trans, ['foo' => 'fo0_b@r'], ['foo' => new StrAlphaUnderscore]);
-        $this->assertTrue($validator->fails());
+        $val = $this->validator(['foo' => 'fo0_b@r'], ['foo' => new StrAlphaUnderscore]);
+        $this->assertTrue($val->fails());
 
         $this->assertSame([
             'foo' => [
                 'The foo may only contain letters, numbers and underscores.',
             ],
-        ], $validator->messages()->messages());
+        ], $val->messages()->messages());
     }
 }
